% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_IRR.R
\name{Possion_IRR}
\alias{Possion_IRR}
\title{Variable Effect IRR(risk ratio/risk ration Interpretation)
This function returns risk ratio coefficient for variable
with 95% confidence interval and Pr(>|z|)}
\usage{
Possion_IRR(possion_model)
}
\arguments{
\item{possion_model}{A fitted possion regression model using glm}
}
\value{
A dataframe that contains the IRR, z value, p-value, and 95% confidence interval for the IRR.
}
\description{
Variable Effect IRR(risk ratio/risk ration Interpretation)
This function returns risk ratio coefficient for variable
with 95% confidence interval and Pr(>|z|)
}
