% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_Prob.R
\name{Predict_Prob}
\alias{Predict_Prob}
\title{Predict Probability of outcome variable
This function returns the probability of the outcome variable
at the value of input under poisson distribution
Note that default is a one-tailed function}
\usage{
Predict_Prob(value, lambda)
}
\arguments{
\item{value}{input expected value of outcome variable}

\item{lambda}{input vector space of predicted lambda}
}
\value{
return a vector space of predicted probability according to

poisson distribution
}
\description{
Predict Probability of outcome variable
This function returns the probability of the outcome variable
at the value of input under poisson distribution
Note that default is a one-tailed function
}
