% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_Prob_Negbinom.R
\name{Predict_Prob_Negbinom}
\alias{Predict_Prob_Negbinom}
\title{Predict Probability of outcome variable
This function returns the probability of the outcome variable
at the value of input under negative binomial distribution
Note that default is a one-tailed function}
\usage{
Predict_Prob_Negbinom(value, theta, lambda)
}
\arguments{
\item{value}{input expected value of outcome variable}

\item{theta}{input the scaling parameter of the std error.}

\item{lambda}{input vector space of predicted lambda}
}
\value{
return a vector space of predicted probability according to

negative binomial distribution
}
\description{
Predict Probability of outcome variable
This function returns the probability of the outcome variable
at the value of input under negative binomial distribution
Note that default is a one-tailed function
}
