% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_Lambda.R
\name{Preidct_Lambda}
\alias{Preidct_Lambda}
\title{Poisson lambda value prediction
This function returns the predicted lamda value of each variable given
a dataframe and the possion regression model}
\usage{
Preidct_Lambda(df, poisson_model)
}
\arguments{
\item{df}{input dataframe of the original dataset}

\item{poisson_model}{poisson_model input fitted poisson regression model using glm}
}
\value{
a vector space of predicted value lambda
}
\description{
Poisson lambda value prediction
This function returns the predicted lamda value of each variable given
a dataframe and the possion regression model
}
